fix: Link의 순환참조 오류 해결
