+++ frontend/src/App.tsx
@@ -3,18 +3,18 @@ import { useContext } from "react";
 import CssBaseline from "@mui/material/CssBaseline";
 import { ThemeProvider } from "@mui/material/styles";
 
-import { Link, RouterContext } from "router/router";
+import RouterContext from "router/Contexts";
+import { pages } from "router/pages";
 import theme from "styles/theme";
 
 const App = (): JSX.Element => {
-  const page = useContext(RouterContext);
+  const { page } = useContext(RouterContext);
 
   return (
     <div className="App">
       <ThemeProvider theme={theme}>
         <CssBaseline />
-        {page}
-        <Link to="searchResult">hi</Link>
+        {pages[page]}
       </ThemeProvider>
     </div>
   );
