+++ frontend/src/router/Contexts.ts
@@ -1,0 +1,11 @@
+import { createContext } from "react";
+
+export default createContext<RouterContextType>({
+  page: "index",
+  setPage: () => {},
+});
+
+interface RouterContextType {
+  page: string;
+  setPage: React.Dispatch<React.SetStateAction<string>>;
+}
