+++ frontend/src/router/router.tsx
@@ -1,44 +1,9 @@
-import { createContext, useContext, useState } from "react";
+import { useMemo, useState } from "react";
 
-import { pages, LinkPath } from "./pages";
+import RouterContext from "./Contexts";
+import { pages } from "./pages";
 
-const { history, location } = window;
-
-export const RouterContext = createContext<React.ReactNode>(pages.index);
-
-const SetPageContext = createContext<SetPageContextType>(() => {});
-type SetPageContextType = React.Dispatch<
-  React.SetStateAction<React.ReactNode>
-> | null;
-
-const pushHistory = ({ path, state }: PushHistoryProps): void => {
-  // history.pushState(state, title, url)
-  history.pushState(state, path, `${location.origin}${path}`);
-};
-
-const Link = ({
-  to,
-  params = {},
-  children,
-  onClick,
-}: LinkProps): JSX.Element => {
-  const setPage = useContext(SetPageContext);
-
-  const href = `/${to}`;
-
-  const handleClick = (e: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
-    e.preventDefault();
-    onClick?.();
-    pushHistory({ path: href, state: params });
-    setPage?.(pages[to]);
-  };
-
-  return (
-    <a href={href} onClick={handleClick}>
-      {children}
-    </a>
-  );
-};
+const { location } = window;
 
 const FIRST_INDEX = 0;
 const FIRST_SLASH_COUNT = 1;
@@ -53,52 +18,30 @@ const getCurrentPath = () => {
 const Router = ({ children }: RouterProps): React.ReactElement => {
   const currentPath = getCurrentPath();
 
-  const [page, setPage] = useState(pages[currentPath] || pages.notFound);
+  const [page, setPage] = useState(pages[currentPath] ? "index" : "notFound");
 
   onpopstate = (/* e: PopStateEvent */) => {
     const poppedPath = getCurrentPath();
     // TODO: e.state 이용하여 뒤로가기 시 검색결과
 
     if (!pages[poppedPath]) {
-      setPage(pages.notFound);
+      setPage("notFound");
       return;
     }
 
-    setPage(pages[poppedPath]);
+    setPage(poppedPath);
   };
 
   return (
-    <RouterContext.Provider value={page}>
-      <SetPageContext.Provider value={setPage}>
-        {children}
-      </SetPageContext.Provider>
+    <RouterContext.Provider
+      value={useMemo(() => ({ page, setPage }), [page, setPage])}
+    >
+      {children}
     </RouterContext.Provider>
   );
 };
-
-interface LinkProps {
-  to: LinkPath;
-  children: React.ReactNode;
-  params?: { [key: LinkPath]: string } | {}; // TODO: key 값 추후에 지정하기
-  onClick?: () => void;
-}
-
-interface PushHistoryProps {
-  path: string;
-  state?:
-    | {
-        [key: string]: string;
-      }
-    | {};
-}
-
-// interface RouterContextType {
-//   page: React.ReactNode;
-//   setPage: React.Dispatch<React.SetStateAction<React.ReactNode>>;
-// }
-
 interface RouterProps {
   children: React.ReactNode;
 }
 
-export { Router, Link };
+export default Router;
